requirements.txt
streamlit
beautifulsoup4
requests
pdfplumber
python-dotenv
openai
